requirements.txt
fastapi
uvicorn
pillow
python-pptx
reportlab
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
multidict==6.1.0
numpy==2.2.2
openai==1.63.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tqdm==4.67.1
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
XlsxWriter==3.2.2
yarl==1.18.3
